+++ NEWS.md
@@ -13,6 +13,7 @@
 * Add resumable uploads to `gcs_upload` to allow uploads over 5MB, limit now 5TB
 * Add retry uploads via `gcs_retry_upload`
 * Add delete object via `gcs_delete_object`
+* Add `gcs_source` to source .R files from GCS
 
 # googleCloudStorageR 0.1.0